Docs (api) : add documentation for sign-up api
* Add MemberDocumentationTest.signUp
