--- src/main/java/com/palmseung/keyword/KeywordConstant.java
@@ -1,9 +1,0 @@
-package com.palmseung.keyword;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class KeywordConstant {
-    public static final String BASE_URI_KEYWORD_API = "/api/keywords";
-}
