--- src/main/java/com/palmseung/keyword/domain/KeywordRepository.java
@@ -1,11 +1,0 @@
-package com.palmseung.keyword.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import javax.transaction.Transactional;
-import java.util.Optional;
-
-@Transactional
-public interface KeywordRepository extends JpaRepository<Keyword, Long> {
-    Optional<Keyword> findByKeyword(String keyword);
-}
