+++ src/main/java/com/palmseung/keywords/KeywordConstant.java
@@ -1,0 +1,9 @@
+package com.palmseung.keywords;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class KeywordConstant {
+    public static final String BASE_URI_KEYWORD_API = "/api/keywords";
+}
