+++ src/main/java/com/palmseung/keywords/domain/KeywordRepository.java
@@ -1,0 +1,11 @@
+package com.palmseung.keywords.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import javax.transaction.Transactional;
+import java.util.Optional;
+
+@Transactional
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
+    Optional<Keyword> findByKeyword(String keyword);
+}
