+++ src/main/java/com/palmseung/keywords/domain/MyKeywordRepository.java
@@ -1,0 +1,11 @@
+package com.palmseung.keywords.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import javax.transaction.Transactional;
+import java.util.List;
+
+@Transactional
+public interface MyKeywordRepository extends JpaRepository<MyKeyword, Long> {
+    List<MyKeyword> findAllByMemberId(Long memberId);
+}
