--- src/main/java/com/palmseung/member/controller/ApiMemberController.java
@@ -1,77 +1,0 @@
-package com.palmseung.member.controller;
-
-import com.palmseung.member.MemberConstant;
-import com.palmseung.member.domain.Member;
-import com.palmseung.member.dto.*;
-import com.palmseung.member.service.MemberService;
-import com.palmseung.support.jwt.JwtTokenProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping(value = MemberConstant.BASE_URI_MEMBER_API)
-public class ApiMemberController {
-    private final MemberService memberService;
-    private final JwtTokenProvider jwtTokenProvider;
-
-    @PostMapping()
-    public ResponseEntity<MemberResponseView> create(@RequestBody CreateMemberRequestView request) {
-        Member createdMember = memberService.create(request);
-        MemberResponseView response = MemberResponseView.of(createdMember);
-
-        return ResponseEntity
-                .created(URI.create(MemberConstant.BASE_URI_MEMBER_API + "/" + createdMember.getId()))
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(response);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<MemberResponseView> deleteById(@PathVariable Long id) {
-        memberService.delete(memberService.findById(id));
-
-        return ResponseEntity
-                .noContent()
-                .build();
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<LoginResponseView> login(@RequestBody LoginRequestView request) {
-        Member member = memberService.login(request.getEmail(), request.getPassword());
-        String token = jwtTokenProvider.createToken(member.getEmail());
-
-        return ResponseEntity
-                .created(URI.create("/oauth/token"))
-                .body(LoginResponseView.of(token));
-    }
-
-    @GetMapping("/my-info/{id}")
-    public ResponseEntity<MemberResponseView> retrieveMyInfo(@PathVariable Long id) {
-        Member member = memberService.findById(id);
-
-        return ResponseEntity
-                .ok()
-                .body(MemberResponseView.of(member));
-    }
-
-    @PutMapping("/my-info/{id}")
-    public ResponseEntity<MemberResponseView> updateMyInfo(@RequestBody UpdateMemberRequestView requestView) {
-        Member updatedMember
-                = memberService.updateInfo(getLoginUser(), Member.of(requestView));
-
-        return ResponseEntity
-                .ok()
-                .body(MemberResponseView.of(updatedMember));
-    }
-
-    private Member getLoginUser() {
-        return (Member) SecurityContextHolder.getContext()
-                .getAuthentication()
-                .getPrincipal();
-    }
-}
