--- src/main/java/com/palmseung/member/domain/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.palmseung.member.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import javax.transaction.Transactional;
-import java.util.Optional;
-
-@Transactional
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
