+++ src/main/java/com/palmseung/members/domain/MemberRepository.java
@@ -1,0 +1,11 @@
+package com.palmseung.members.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import javax.transaction.Transactional;
+import java.util.Optional;
+
+@Transactional
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+}
