+++ src/main/java/com/palmseung/members/dto/CreateMemberRequestView.java
@@ -1,0 +1,43 @@
+package com.palmseung.members.dto;
+
+import com.palmseung.members.domain.Member;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class CreateMemberRequestView {
+    private String email;
+    private String name;
+    private String password;
+
+    public CreateMemberRequestView() {
+    }
+
+    @Builder
+    public CreateMemberRequestView(String email, String name, String password) {
+        this.email = email;
+        this.name = name;
+        this.password = password;
+    }
+
+    public static CreateMemberRequestView of(Member member) {
+        return CreateMemberRequestView.builder()
+                .email(member.getEmail())
+                .name(member.getName())
+                .password(member.getPassword())
+                .build();
+    }
+
+    public CreateMemberRequestView changePassword(String encodedPassword) {
+        this.password = encodedPassword;
+        return this;
+    }
+
+    public Member toEntity() {
+        return Member.builder()
+                .email(email)
+                .name(name)
+                .password(password)
+                .build();
+    }
+}
