+++ src/main/java/com/palmseung/members/dto/LoginRequestView.java
@@ -1,0 +1,23 @@
+package com.palmseung.members.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Getter
+public class LoginRequestView {
+    private String email;
+    private String password;
+    private String accessToken;
+    private String tokenType;
+
+    @Builder
+    public LoginRequestView(String email, String password,
+                            String accessToken, String tokenType) {
+        this.email = email;
+        this.password = password;
+        this.accessToken = accessToken;
+        this.tokenType = tokenType;
+    }
+}
