+++ src/test/java/com/palmseung/BaseDocumentationTest.java
@@ -1,0 +1,33 @@
+package com.palmseung;
+
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.palmseung.members.domain.MemberRepository;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
+@AutoConfigureRestDocs
+@AutoConfigureMockMvc
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
+@Import(RestDocsConfig.class)
+public class BaseDocumentationTest {
+    @Autowired
+    public ObjectMapper objectMapper;
+
+    @Autowired
+    public MemberRepository memberRepository;
+
+    @AfterEach
+    public void cleanDatabase() {
+        memberRepository.deleteAll();
+    }
+}
