+++ src/test/java/com/palmseung/RestDocsConfig.java
@@ -1,0 +1,23 @@
+package com.palmseung;
+
+import org.springframework.boot.test.autoconfigure.restdocs.RestDocsMockMvcConfigurationCustomizer;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.restdocs.mockmvc.MockMvcRestDocumentationConfigurer;
+
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+
+@TestConfiguration
+public class RestDocsConfig {
+    @Bean
+    public RestDocsMockMvcConfigurationCustomizer restDocsMockMvcConfigurationCustomizer() {
+        return new RestDocsMockMvcConfigurationCustomizer() {
+            @Override
+            public void customize(MockMvcRestDocumentationConfigurer configurer) {
+                configurer.operationPreprocessors()
+                        .withRequestDefaults(prettyPrint())
+                        .withResponseDefaults(prettyPrint());
+            }
+        };
+    }
+}
